Resource: redefine types, create route for resource on backend
