+++ packages/types/resource.ts
@@ -1,0 +1,9 @@
+export interface MockResource {
+    id: string;
+    title: string;
+    sections: Section[];
+}
+export interface Section {
+    header: string;
+    content: string;
+}
